Handle block events in BokehPlot rather than DataBlockBase
Only Bokeh plots dispatch `block-event` -- the sole producer is
`generate_js_callback_single_float_parameter` -- so the listener belongs
with the plot rather than on the generic block wrapper, which also runs for
blocks that have no plot at all.

BokehPlot is also the component that holds the live Bokeh document, so
putting the handler here is what will let an event patch the plot in place
instead of going through the store and forcing a re-embed.

Pure move: the handler body is unchanged, still filtering on block_id and
still calling updateBlockFromServer. It reaches the store directly now
rather than through `this.$store`, and takes the two ids it needs as props.
