+++ webapp/src/components/datablocks/BokehBlock.vue
@@ -25,7 +25,7 @@
 
     <template #plot>
       <div id="bokehPlotContainer" class="limited-width">
-        <BokehPlot :bokeh-plot-data="bokehPlotData" />
+        <BokehPlot :item_id="item_id" :block_id="block_id" :bokeh-plot-data="bokehPlotData" />
       </div>
     </template>
   </DataBlockBase>
