+++ webapp/src/components/datablocks/UVVisBlock.vue
@@ -11,7 +11,7 @@
       />
     </div>
     <div id="bokehPlotContainer">
-      <BokehPlot :bokeh-plot-data="bokehPlotData" />
+      <BokehPlot :item_id="item_id" :block_id="block_id" :bokeh-plot-data="bokehPlotData" />
     </div>
   </DataBlockBase>
 </template>
